update absensi work

+++ database/migrations/2025_10_01_103032_remove_chat_id_from_support_requests_table.php
@@ -11,8 +11,10 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('support_requests', function (Blueprint $table) {
-            $table->dropForeign(['chat_id']);
-            $table->dropColumn('chat_id');
+            if (Schema::hasColumn('support_requests', 'chat_id')) {
+                $table->dropForeign(['chat_id']);
+                $table->dropColumn('chat_id');
+            }
         });
     }
